Increase ULS menu z-index when opened from AC/DC
WikibaseMediaInfo uses UniversalLanguageSelector to implement
monolingual text values, but ULS doesn’t play very nicely with OOUI: it
just inserts its element at the beginning of the body, relying on the
z-index in order to be displayed above the elements that actually come
after it.  Its built-in styles include z-index 1000, but the Vector skin
overrides this, lowering the z-index to just 3 – much lower than the
@z-index-overlay variable, also from Vector, which is used for OOUI
dialogs in that skin. This means that ULS displays beneath AC/DC,
effectively becoming unusable under Vector. (In Monobook, on the other
hand, it works fine, due to less z-index manipulation in that skin.)

Fix that with a custom style rule to increase the z-index again. This is
a pretty ugly hack, so I’m trying to migitage its bad effects: we only
do this when AC/DC is active (adding a new class to the <body> for this
purpose, since we can’t use :has(.acdc-statementsDialog)); we set the
z-index to the lowest number that works for us (102, just one above
Vector’s 101); and we include a custom variable in the ruleset, which
people can see in their browser’s developer tools to find out where this
z-index is coming from, if it happens to disturb anything they’re doing.