Fix browser test for latest Vector 2022 changes
The sidebar has apparently been split, and AC/DC goes with the other
Tools into a separate menu that we need to use a different selector for.

On the plus side, we can simplify the code: we only need to use one
element, and we also don’t need to explicitly wait for it to exist
because WebdriverIO does that implicitly.