Remove wikibase.common ResourceLoader dependency
This was added in commit f4c1b4305a to pull in the styles that hide
language fallback indicators for language variants; for this purpose, it
hasn’t been needed since T345883 [1] last year. More recently, in
Wikibase change Id3241ab944 (commit 2d1985a97a [2]), the wikibase.common
module was completely removed (by yours truly); we could replace it with
wikibase.alltargets or wikibase.desktop if we needed any styles from it,
but I think it should actually be fine to just remove it completely.

This was detected by the browser tests – and I think this might be the
first ever time that having the browser tests actually paid off. It only
took 5½ years! (At least, I don’t remember any specific other time that
the browser tests detected a problem that needed to be fixed in AC/DC
itself, and I didn’t find anything in the git log either.)

[1]: https://phabricator.wikimedia.org/T345883
[2]: https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/1129869