Lower language level from ES2019 to ES2017
ES2017 is natively supported by MediaWiki (on Wikimedia wikis) now,
thanks to T381537; this is close enough to our previous language level
(most importantly, supporting async/await!) that we can restrict our
code to this level without it being too much of a chore.

The single for-await-of loop can be replaced with a callback pattern (we
don’t even need support for breaking from the loop, though that wouldn’t
be too hard to support either by returning a value from the callback).
Object.fromEntries() can be replaced with a manually constructed object.
.flatMap() is a bit more annoying (it’s a shortcut for .map().flat() but
.flat() is also ES2019), but I think the [].concat(...) approach isn’t
too terrible.