Add higher timeouts for searches
Searching for files, properties or items to select can also take a long
time, so introduce a dedicated timeout for this (in a new file, since
it’s also needed in pageobjects/ rather than just specs/) and make it
longer than the 10000 ms (10 s) default timeout. The “can remove
multiple statements” test seemed to be especially susceptible to search
timeouts and should hopefully be more stable now.