Wait for portlet link to be displayed before checking it
This test was occasionally failing with the actual text being the empty
string, and looking at it I realized that it never actually waited for
the portlet link to be displayed – I guess it mostly worked by virtue of
waiting for AC/DC to be installed and the tools menu to be shown. Let’s
see if this helps to make the test more stable.