Support user scripts based on AC/DC better
It’s been brought to my attention that there are some user scripts that
interact with AC/DC, using some fairly rickety hacks (mainly via the
DOM, similar to how the browser tests work, except without even basic
affordances like .waitForDisplayed()). Let’s try to offer a better
interface. (Of course, this comes with absolutely no promises for
stability or anything.)

The centerpiece of this is the new “exports” object. I initially planned
to make this available via the module.exports (which package gadgets can
have), but then I realized that we only assign to the exports after
awaiting several network requests, which makes them unreliable to work
with. We still assign to the module.exports for completeness’ sake, but
the much better way to access them is via the argument to the (already
existing) gadget.acdc.loaded hook.

The only export so far is an openAcdc() function, which does what it
says on the tin. It returns a promise for the WindowInstance (with its
lifecycle hooks) and the StatementsDialog (note that, while this is
async, OOUI will resolve the promise basically instantly, and *not* wait
until the dialog is actually opened), and you can then call further
methods on the dialog; the existing acdcShow and acdcPagePileId URL
parameters are now implemented via this interface. However, I noticed
that trying to interact with the statement widgets after awaiting
openAcdc() didn’t work properly: this is because the StatementsDialog
has to asynchronously fetch the data types of the configured properties,
and can only add the statement widgets afterwards. So the
StatementsDialog now contains a flag which you can use to check if this
initialization is finished, and fires an event when it happens;
openAcdc() turns this, along with the “opened” lifecycle promise of the
window manager, into a new hook, gadget.acdc.opened, which other scripts
can subscribe to (it’s called with the statementsDialog for convenience,
wrapped in an object for future-proofing). As soon as this hook fires,
you can also modify the statements in the statement widgets.

To make modifying the statement widgets more convenient, I also tweaked
addStatementToAddWidget() and addStatementToRemoveWidget() to return the
added widget and also make sure not to re-add it, so that a user script
can call e.g. addStatementToAddWidget( 'P180', 'wikibase-item' ) and
then always work on the “depicts” widget regardless of what the user’s
favorite properties are.

(Don’t ask me why jsdoc/no-undefined-types needs to be taught that the
StatementsDialog and StatementWidget types are defined. You’d think it
could easily find them in the scope…)