Implement two TODOs related to references
WikibaseMediaInfo has supported references for a while now (T230315), so
it was high time that I actually fixed these two TODOs. Though keep in
mind that they were both for edge cases – as far as I can tell, the
common case (adding new statements with references) should’ve already
been working just fine.

The smaller, but more relevant part is adding support for references
when updating an already existing statement: add any reference as a new
reference to the statement, unless it already has the same reference.
That is, you can’t use AC/DC to update an existing reference (add snaks
to it), which should hopefully be a rare case. Previously, the
references would’ve been left unmodified in this case. (And in case
you’re wondering: yes, the check for existing references is necessary,
Wikibase doesn’t filter them out server-side: see [1] for an example.)

The other, larger but less relevant part is just adding an error message
and preventing saving if the user specifies references in the statements
to remove. All this is just copy+pasted from treating qualifiers in the
same way.

[1]: https://commons.wikimedia.beta.wmcloud.org/w/index.php?title=File:ACDC_test_file_1.pdf&diff=prev&oldid=322495