Don’t allow selecting already present properties
AddPropertyWidget internally keeps track of which properties have been
selected already, so that users can’t add a second widget for that
property (it’s filtered out of the search results). However, this was
previously unconnected to the notion of the property widgets which AC/DC
adds by default, so it was possible to get e.g. two “depicts” widgets
(one added by default and one from searching for the property manually).
Prevent this by passing (a copy of) the properties we’re going to start
out with into the AddPropertyWidget. (If we supported removing property
widgets, we’d want to call .onStatementPanelRemoved() on the
AddPropertyWidget to notify it that the property is eligible for
selection again, but right now that’s not supported.)