Implement error for adding and removing same statement
I suspect when I wrote this TODO, I intended to implement the error as
part of the original “remove” feature (which ended up becoming commit
b39517c3b4). But I forgot, so here I am, instead implementing it some
*checks* what the hell, five years later?!

Anyway. Nothing too complicated here, though it’s a bit different from
the other errors in that it depends on the add and remove widgets
simultaneously, so the check is implemented in a separate method
(whereas the other checks are right inside the 'change' handlers inline
inside addStatementToAddWidget() / addStatementToRemoveWidget()).
Extract a “peek” version of those “add” methods so we can check if they
exist or not without adding the widget if it doesn’t exist yet, which
could be useful to others as well. Note that installStyles() doesn’t
need to be changed because it already included the selector for the
.acdc-statementsDialog__statementWidget--statement-to-add-and-remove
class.